+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## Unreleased
 ### Added
 - frontend translating with ttag
+- frontend validation for portfolio name
 
 ## [1.0.2] - 19.01.2021
 ### Added